feat: :wrench: acrescentado configuracao para rodar testes e2e separados dos testes unitarios
