+++ package.json
@@ -6,9 +6,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "NODE_ENV=test jest --detectOpenHandles",
-    "test:clearCache": "NODE_ENV=test jest --clearCache",
     "dev": "tsnd --respawn --transpile-only -r tsconfig-paths/register src/shared/infra/http/index.ts",
+    "test": "jest --config jest.units.config.ts",
+    "test:debug": "jest --config jest.units.config.ts --debug",
+    "test:e2e": "NODE_ENV=test jest --config jest.e2e.config.ts",
+    "test:e2e:debug": "NODE_ENV=test jest --config jest.e2e.config.ts --debug",
     "typeorm": "ts-node-dev -r tsconfig-paths/register ./node_modules/typeorm/cli.js",
     "migration:create": "cd src/shared/infra/typeorm/migrations/ && npx typeorm migration:create",
     "migration:up": "./node_modules/.bin/ts-node ./node_modules/.bin/typeorm migration:run -d src/shared/infra/typeorm/index.ts",
