+++ src/modules/person/tests/mocks/ClientMock.ts
@@ -1,4 +1,4 @@
-import { ClientRepository } from '../repositories/ClientRepository';
+import { ClientRepository } from '@modules/person/repositories/ClientRepository';
 
 export const clientRepositoryMock: jest.Mocked<ClientRepository> = {
   findAll: jest.fn(),
