+++ src/modules/user/tests/mocks/HashProviderMock.ts
@@ -1,4 +1,4 @@
-import HashProvider from '../providers/HashProvider/models/HashProvider';
+import HashProvider from '@modules/user/providers/HashProvider/models/HashProvider';
 
 export const hashProviderMock: jest.Mocked<HashProvider> = {
   generateHash: jest.fn(),
