+++ src/shared/infra/typeorm/seeds/clients.ts
@@ -1,4 +1,3 @@
-import { randomUUID } from 'crypto';
 import { uniqueNamesGenerator, names } from 'unique-names-generator';
 
 import { getDbConnection } from '..';
@@ -12,12 +11,11 @@ async function create() {
 
   const forLoop = async () => {
     for await (const iterator of Array.from(Array(NUMBER_OF_COSTUMERS).keys())) {
-      const id = randomUUID();
       const randomName = uniqueNamesGenerator({
         dictionaries: [names]
       })
 
-      await connection.query(`INSERT INTO CLIENT (id, name) VALUES ('${id}', '${randomName}')`);
+      await connection.query(`INSERT INTO CLIENT (name) VALUES ('${randomName}')`);
     }
   }
